cmake: work around a mysterious macOS dyld problem.
For some reason, there appears to be no way to build Wireshark on macOS
Ventura with Xcode 15 and have the resulting program run without getting
an error due to failing to find pcap_createsrcstr() at startup.

In particular, there appears to be no way to use __builtin_available()
to protect accesses to the routines that showed up in Sonoma, so that
the run-time linker doesn't fail if the routine in question isn't
present.  Perhaps it requires more compiler command-line arguments.

So, instead, only check for pcap_open() if 1) this isn't macOS or 2) the
pcap library pathname doesn't begin with /usr/lib/ or
/Application/Xcode.app/, i.e. if it's not the system libpcap.  That way,
we won't every try to use those APIs for a built with the system
libpcap.

Fixes #19436.