INAP: Fix dissection of local code in ExtensionField
Previously when a message contains an Extension Field with a local
code the operation name corresponding to this local code was added to
the Info column. This is incorrect as extensions local codes are not
operations.

This change is similar to how Extension Field is handled in the CAMEL
protocol.