macos-setup: update to zstd 1.5.5. [skip ci]
It has a number of performance improvements and, even more important...

...it appears to let our settings for the minimum version to which code
can be deployed on macOS (and for the SDK to use) pass through to the
entire build process, unlike 1.4.5, which somehow seems to have
prevented that, which is why #19405 kept popping up.

Thanks to @tuexen for finding out that 1.5.5 seemed to finally get a
build of the libzstd lbirary that doesn't use Exciting New Run-Time
Linker Features that aren't supported by the run-time linker on older
versions of macOS for which we support Wireshark.