TNS: Handle connect data in subsequent packet and data descriptor type
In Oracle TNS, when the Connect Data length is long (> 221 bytes,
apparently) it is not carried in the Connect PDU at all but in a
Data PDU immediately following. Notice when that happens, and avoid
marking the Connect packet as malformed. Mark the subsequent Data
packets as containing Connect Data and process it as a string instead
of just bytes.

Handle the new data descriptor TNS type mentioned elsewhere, which
has (LOB/FILE) data immediately following the PDU and not counted
in the PDU length. Since this does not have another PDU header,
determine it in the get pdu length functions, returning
ONE_MORE_SEGMENT in the cases where we only have the fixed header
but not enough bytes to read the LOB/FILE data following.

This makes it possible to view the images in the attached capture
with "Show Packet Bytes".

Fix #15727