Qt: Fix Export Packet Dissections Help button, connections on non-Windows
We have to call findChild to find the QDialogButtonBox after setting
the option to use the Qt dialog instead of the native dialog, not
before. Otherwise, we might not find the QDialogButtonBox.

Fixes adding the Help button to the Export Packet Dissections
dialog, and enabling the validity checks that disable the
save button when, e.g. an illegal range is entered.