Fix saving recently used capture filters
The recently used capture filters list was organized so that
the newest filters were last, but when writing to recent_common
we start with the beginning of the list and write up to the max.
This meant that only the first 20 entries ever used for a profile
would be written, after which the list would be frozen.

Fix that. The best way to do that is to keep the most recently
used filter first, in presentation order, by prepending each
filter. That way we can AddItems to the combo boxes and write
to the file in the natural way. In order to do so, we have to
reverse the list after reading all the entries from the recent_common
file on startup. This also has better performance, as it's O(N) to
prepend a bunch of list items and them reverse them, but O(N^2)
to add N list items to the end of a linked-list (that doesn't have
a pointer to the end, unlike a GQueue.)

Fix #12918