Qt: Don't try to compile capture filters for unknown dlts
If an interface's active_dlt is set to -1 (an invalid value, used
when permissions are bad, or set by CaptureOptionsDialog for
"link-layer types libpcap/WinPcap/Npcap doesn't know about"),
don't try to compile and check capture filters. Give the same
yellow Deprecated background and warning about being unable to
check the capture filter as for the DLT_USER types, instead of a
cryptic error message.

Fix #19480