docs: Alter FollowStreamSection anchor
According to the Asciidoctor docs, we can assign multiple anchors
to a section using this method, which still uses the modern shorthand
syntax for the main anchor (the one that's been used for the last five
years). We have to use the double square brackets for the additional
anchor, though.

    https://docs.asciidoctor.org/asciidoc/latest/attributes/id/#add-additional-anchors-to-a-section

Related to #17982