Qt: Store active remote hosts in a preference file
Store the active remote hosts in a preference file.
Use JSON because in addition to each host, we later may want
to save the per-interface settings (but the interfaces
available on a host could change without the host information
changing). Hosts and interfaces together have a tree-like data
structure that doesn't work with a UAT or other table.

Still TODO: password handling (cf. #17949 and extcap.), moving
this into a Model instead of a TreeWidget, getting rid of
remote_interface_dialog and handling all of this in
ManageInterfacesDialog through the model.

Fix #17484. Fix #8557.