Qt: Save the main window QSplitter states and restore them
This is more accurate, because it accounts for the AccordionFrame
being open or closed, than looking at the visible pixel count.
Keep saving the older method that was also used for the GTK UI,
and fallback to it if this fails. That's also useful for a first
run and for people who switch back and forth with older versions.

Fix #19361