Qt: Remember what monitor we last ran on when maximizing
Use Qt's save and restoreGeometry functions when trying to restore
to a maximized window state. Those functions do a better job
tracking on what screen we were maximized. It also will cause
the non-maximized size to be saved and restored even when starting
out maximized, insted of reverting to the default size when leaving
maximization.

Maybe in the future we would want to use these functions for all the
recent geometry handling. We would lose our current pref granularity
allowing size, position, and maximized status to each be saved or not.
(Do we need that?)

There are still some edge cases:
https://bugreports.qt.io/browse/QTBUG-77385
https://bugreports.qt.io/browse/QTBUG-70721
some of which are resolved in the latest Qt, but it's better than
the current handling.

Fix #12389