rdp: various fix for the dynamic channel
The packet reassembly was not always done correctly and was sometime making wireshark
to segfault. The patch reworks packet reassembly, making it a little simpler. It also
tracks first and last packet of reassembly, so if needed we could add links to these frame
in the future.
Also the progress information used to also be erronous in some intermediate packets, the
patch fixes that.