Zigbee make the formatting of Zigbee UTC Time more readable and usable
zbee_zcl.attr.utc is a string representation of the time. This time is based on an uncommon epoch.
Someimes the numerical value for UTCTime is convenient (e.g. sorting by numbers rather than
string represenations of time). This numerical value is now both displayed in the UTC Time field
and a second filter field utc_raw has been added for when the numerical value is needed,
for example when using t-shark.

Thanks to Cole Wu <colewu9712@gmail.com> for the original implementation and support validating
this update.