mgcp: Fix a buffer overrun
Don't assume that a string in packet data necessarily stops
in 256 octets; that may be the specifiction, but it's not true
for fuzzed data.
Use tvb_ functions when possible to avoid issues like this.

Fun Fact: I had to rewrite the loop to put getting the tempchar
in the body, because having everything in the for loop declaration
caused gcc to optimize the loop out of existence.

Fixup b28619576becb176926858ac32ec3fdd03cda019

Related to #19501