Qt: Save GeometryStateDialog splitter states
Save the state of QSplitters in GeometryStateDialogs to
the recent settings. (Per-profile, because the desired splitter
setting can depend on the preference for the packet dialog, or
the number of IO graphs loted for the IO Graph dialog, etc., which
are stored per-profile.)

Don't restore the splitter state for the Preferences Dialog, because
the calculations done to set the minimize size of the protocol
column make it difficult to shrink the splitter below the size of
the longest protocol once it gets set larger.

Fix #18121