TECMP: Make ASAM CMP detection backward compatible (Bugfix)
TECMP 1.7 states that the first byte of the Device ID is always 0x00, so
we can detect ASAM CMP on this. The older TECMP versions allowed that
byte to 0xff for user-defined Device IDs. Since traces with TECMP 1.6
and before still exist, the ASAM CMP Detection is updated to allow the
first byte to be 0xff in TECMP too.