Resolved Addresses: Only show ethers actually used in the capture file
Change the hashether status flags to be more consistent with the others.
In the Resolved Addresses dialog, only show addresses that are
actually present in the capture file and resolved, not all entries
in an ethers or wka file. (Entries in the wka file with a /48 mask
are added to the ethers hashtable.)

Part of #13857