Qt: Add Copy, Save As to Resolved Addresses dialog
Add basic Copy to clipboard and Save As functionality to the
Resolved Addresses dialog. This saves in text formats compatible
with hosts(5), ethers(5), and services(5), with tabs between the
columns, along with CSV and JSON.

This copies whatever is currently visible in the dialog.

More file formats could be added later, or a Context Window
pop-up; the latter could allow Copying or Saving only the selected
items.

Fix #16419