Qt: Replace plugin column "Path" with "Scope" in About dialog
Instead of using a Path field, that is really long and contains
a lot of redundant information, use a scope field that indicates
if the plugin is a personal plugin (loaded from the personal plugins
folder) or a global plugin (loaded from the global plugins folder).

Together with the plugin name this allows to construct the path
completely. By right-clicking on the row it is possible to open the file
browser in the correct plugin path and the personal/global plugin
folder path itself is displayed elsewhere.

Extcap and Lua scripts need to be enhanced to provide this
information too.